Moved pretty formatter test under utils test class
