+++ src/test/java/com/amazon/opendistroforelasticsearch/sql/unittest/utils/PrettyFormatterTest.java
@@ -13,7 +13,7 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.sql.unittest;
+package com.amazon.opendistroforelasticsearch.sql.unittest.utils;
 
 import com.amazon.opendistroforelasticsearch.sql.esintgtest.TestUtils;
 import com.amazon.opendistroforelasticsearch.sql.utils.JsonPrettyFormatter;
